requirements.txt
PyQt6>=6.0.0
pyserial>=3.4